JWT authentication added

+++ resources/item.py
@@ -1,7 +1,6 @@
-import uuid
-from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
+from flask_jwt_extended import jwt_required, get_jwt
 from db import db
 from sqlalchemy.exc import SQLAlchemyError
 from models import ItemModel
@@ -9,8 +8,9 @@ from schemas import ItemSchema, ItemUpdateSchema
 
 blp = Blueprint("items", __name__, description="Operations on items")
 
-@blp.route("/item/<string:item_id>")
+@blp.route("/item/<int:item_id>")
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):   
         try:
@@ -19,8 +19,12 @@ class Item(MethodView):
         except SQLAlchemyError:
             abort(500, message="An error occurred while fetching the item.")
 
+    @jwt_required(fresh=True)
     def delete(self, item_id):
         try:
+            jwt = get_jwt()
+            if jwt.get("is_admin", False) is False:
+                abort(401, message="Admin privileges required to delete an item.")
             item = ItemModel.query.get_or_404(item_id)
             db.session.delete(item)
             db.session.commit()
@@ -28,6 +32,7 @@ class Item(MethodView):
         except SQLAlchemyError:
             abort(500, message="An error occurred while deleting the item.")
 
+    @jwt_required(fresh=True)
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
     def put(self, item_data, item_id):
@@ -46,6 +51,7 @@ class Item(MethodView):
 
 @blp.route("/item")
 class ItemList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):
         try:
@@ -53,6 +59,7 @@ class ItemList(MethodView):
         except SQLAlchemyError:
             abort(500, message="An error occurred while fetching items.")
     
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, request_data):
@@ -61,5 +68,5 @@ class ItemList(MethodView):
             db.session.add(item)
             db.session.commit()
             return item
-        except SQLAlchemyError:
-            abort(500, message="An error occurred while inserting the item.")
+        except SQLAlchemyError as e:
+            abort(500, message="An error occurred while inserting the item. error: {}".format(str(e)))
